优化独立运行所需的兼容类
