+++ composer.json
@@ -26,7 +26,9 @@
         "psr-4": {
             "think\\": "src"
         },
-        "files": []
+        "files": [
+            "stubs/load_stubs.php"
+        ]
     },
     "autoload-dev": {
         "psr-4": {
